fix: keep track of message contexts
The context of the Messages in the original PO file must be saved, if
present. If the Message has a context, then it is necessary to specify
it when retrieving it from the Catalog with the `get` method, otherwise
it returns None. This is what caused the crash in the previous versions
of this script.